UDAF cleanup.
